requirements/doc.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
aiohttp==3.6.2            # via -r requirements/test.txt, github.py, pytest-aiohttp
alabaster==0.7.12         # via sphinx
async-timeout==3.0.1      # via -r requirements/test.txt, aiohttp
attrs==19.3.0             # via -r requirements/test.txt, aiohttp, pytest
babel==2.8.0              # via sphinx
bleach==3.1.5             # via readme-renderer
certifi==2020.6.20        # via requests
chardet==3.0.4            # via -r requirements/test.txt, aiohttp, doc8, requests
coverage==5.2             # via -r requirements/test.txt, pytest-cov
doc8==0.8.1               # via -r requirements/doc.in
docutils==0.16            # via doc8, readme-renderer, restructuredtext-lint, sphinx
edx-sphinx-theme==1.5.0   # via -r requirements/doc.in
gitdb==4.0.5              # via -r requirements/test.txt, gitpython
github.py==0.5.0          # via -r requirements/test.txt
gitpython==3.1.7          # via -r requirements/test.txt
idna-ssl==1.1.0           # via -r requirements/test.txt, aiohttp
idna==2.10                # via -r requirements/test.txt, idna-ssl, requests, yarl
imagesize==1.2.0          # via sphinx
importlib-metadata==1.7.0  # via -r requirements/test.txt, pluggy, pytest
jinja2==2.11.2            # via sphinx
markupsafe==1.1.1         # via jinja2
more-itertools==8.4.0     # via -r requirements/test.txt, pytest
multidict==4.7.6          # via -r requirements/test.txt, aiohttp, yarl
packaging==20.4           # via -r requirements/test.txt, bleach, pytest, sphinx
pathlib2==2.3.5           # via -r requirements/test.txt, pytest
pbr==5.4.5                # via stevedore
pluggy==0.13.1            # via -r requirements/test.txt, pytest
py==1.9.0                 # via -r requirements/test.txt, pytest
pygments==2.6.1           # via doc8, readme-renderer, sphinx
pyparsing==2.4.7          # via -r requirements/test.txt, packaging
pytest-aiohttp==0.3.0     # via -r requirements/test.txt
pytest-cov==2.10.0        # via -r requirements/test.txt
pytest==5.4.3             # via -r requirements/test.txt, pytest-aiohttp, pytest-cov
pytz==2020.1              # via babel
pyyaml==5.3.1             # via -r requirements/test.txt
readme-renderer==26.0     # via -r requirements/doc.in
requests==2.24.0          # via sphinx
restructuredtext-lint==1.3.1  # via doc8
six==1.15.0               # via -r requirements/test.txt, bleach, doc8, edx-sphinx-theme, packaging, pathlib2, readme-renderer, stevedore
smmap==3.0.4              # via -r requirements/test.txt, gitdb
snowballstemmer==2.0.0    # via sphinx
sphinx==3.1.2             # via -r requirements/doc.in, edx-sphinx-theme
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
stevedore==1.32.0         # via doc8
typing-extensions==3.7.4.2  # via -r requirements/test.txt, aiohttp
urllib3==1.25.9           # via requests
wcwidth==0.2.5            # via -r requirements/test.txt, pytest
webencodings==0.5.1       # via bleach
yarl==1.4.2               # via -r requirements/test.txt, aiohttp
zipp==1.2.0               # via -r requirements/test.txt, importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
